Move getRlpHeaderSize to common/util.ts
