+++ ethereum_history_api/oracles/src/noir/oracles/accountOracle/encode.test.ts
@@ -4,14 +4,7 @@ import accountAsFields from './fixtures/accountAsFields.json';
 import stateProofAsFields from './fixtures/stateProofAsFields.json';
 import storageProofAsFields from './fixtures/storageProofAsFields.json';
 import { ForeignCallOutput } from '@noir-lang/noir_js';
-import {
-  RLP_SHORT_ENTITY_MAX_LEN,
-  encodeAccount,
-  encodeStateProof,
-  encodeStorageProof,
-  getRlpHeaderSize
-} from './encode.js';
-import { ADDRESS_LEN, BYTES32_LEN } from '../common/const.js';
+import { encodeAccount, encodeStateProof, encodeStorageProof } from './encode.js';
 
 describe('AccountOracle encode', () => {
   describe('encodeAccount', () => {
@@ -36,36 +29,6 @@ describe('AccountOracle encode', () => {
         serializeStorageProof(storageProofAsFields)
       );
     });
-  });
-});
-
-describe('getRlpHeaderSize', () => {
-  it('should return 1 for short strings', () => {
-    expect(getRlpHeaderSize(0)).toBe(1);
-    expect(getRlpHeaderSize(1)).toBe(1);
-    expect(getRlpHeaderSize(ADDRESS_LEN)).toBe(1);
-    expect(getRlpHeaderSize(BYTES32_LEN)).toBe(1);
-    expect(getRlpHeaderSize(RLP_SHORT_ENTITY_MAX_LEN)).toBe(1);
-  });
-
-  it('should return 2 for long strings where length len is 1', () => {
-    expect(getRlpHeaderSize(RLP_SHORT_ENTITY_MAX_LEN + 1)).toBe(2);
-    expect(getRlpHeaderSize(2 ** 8 - 1)).toBe(2);
-  });
-
-  it('should return 3 for long strings where length len is 2', () => {
-    expect(getRlpHeaderSize(256)).toBe(3);
-    expect(getRlpHeaderSize(512)).toBe(3);
-    expect(getRlpHeaderSize(2 ** 16 - 1)).toBe(3);
-  });
-
-  it('should return 4 for long strings where length len is 3', () => {
-    expect(getRlpHeaderSize(2 ** 16)).toBe(4);
-    expect(getRlpHeaderSize(100_000)).toBe(4);
-  });
-
-  it('should return 5 for long strings where length len is 4', () => {
-    expect(getRlpHeaderSize(100_000_000)).toBe(5);
   });
 });
