+++ ethereum_history_api/oracles/src/noir/oracles/common/util.ts
@@ -1,0 +1,11 @@
+import { BYTE_HEX_LEN } from '../../../util/const.js';
+
+export const RLP_SHORT_ENTITY_MAX_LEN = 55;
+
+export function getRlpHeaderSize(strLen: number): number {
+  if (strLen <= RLP_SHORT_ENTITY_MAX_LEN) {
+    return 1;
+  } else {
+    return 1 + Math.ceil(strLen.toString(16).length / BYTE_HEX_LEN);
+  }
+}
